requirements.txt
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
distro==1.9.0
Flask==3.1.0
flask-cors==5.0.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jwt==1.3.1
MarkupSafe==3.0.2
openai==1.70.0
packaging==24.2
pluggy==1.5.0
psutil==7.0.0
pycparser==2.22
pydantic==2.11.2
pydantic_core==2.33.1
pytest==8.3.5
PyYAML==6.0.2
requests==2.32.3
setuptools==75.8.0
sniffio==1.3.1
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
wheel==0.45.1
